+++ ui/dashboards/src/components/DeletePanelGroupDialog/DeletePanelGroupDialog.tsx
