+++ ui/dashboards/src/components/index.ts
@@ -14,6 +14,7 @@
 export * from './Dashboard';
 export * from './DashboardToolbar';
 export * from './DeletePanelDialog';
+export * from './DeletePanelGroupDialog';
 export * from './GridLayout';
 export * from './Panel';
 export * from './PanelDrawer';
